61_adapter_and_holder_separated
